move instance-related type errors to InstanceError
